Move comment-related printer methods to separate directory
