+++ src/language-js/comments/handle-comments.js
@@ -9,32 +9,25 @@ import {
   addDanglingComment,
   getNextNonSpaceNonCommentCharacterIndex,
   isNonEmptyArray,
-} from "../common/util.js";
+} from "../../common/util.js";
 import {
   getFunctionParameters,
   isPrettierIgnoreComment,
-  isJsxNode,
-  hasFlowShorthandAnnotationComment,
-  hasFlowAnnotationComment,
-  hasIgnoreComment,
   isCallLikeExpression,
   getCallArguments,
   isCallExpression,
   isMemberExpression,
   isObjectProperty,
   isLineComment,
-  getComments,
-  CommentCheckFlags,
   markerForIfWithoutBlockAndSameLineComment,
-} from "./utils/index.js";
-import { locStart, locEnd } from "./loc.js";
-import isBlockComment from "./utils/is-block-comment.js";
-import isTypeCastComment from "./utils/is-type-cast-comment.js";
+} from "../utils/index.js";
+import { locStart, locEnd } from "../loc.js";
+import isBlockComment from "../utils/is-block-comment.js";
+import isTypeCastComment from "../utils/is-type-cast-comment.js";
 
 /**
- * @typedef {import("./types/estree").Node} Node
- * @typedef {import("./types/estree").Comment} Comment
- * @typedef {import("../common/ast-path.js").default} AstPath
+ * @typedef {import("../types/estree").Node} Node
+ * @typedef {import("../types/estree").Comment} Comment
  *
  * @typedef {Object} CommentContext
  * @property {Comment} comment
@@ -918,72 +911,12 @@ function isRealFunctionLikeNode(node) {
   );
 }
 
-/**
- * @param {any} node
- * @returns {Node[] | void}
- */
-function getCommentChildNodes(node, options) {
-  // Prevent attaching comments to FunctionExpression in this case:
-  //     class Foo {
-  //       bar() // comment
-  //       {
-  //         baz();
-  //       }
-  //     }
-  if (
-    (options.parser === "typescript" ||
-      options.parser === "flow" ||
-      options.parser === "acorn" ||
-      options.parser === "espree" ||
-      options.parser === "meriyah" ||
-      options.parser === "__babel_estree") &&
-    node.type === "MethodDefinition" &&
-    node.value &&
-    node.value.type === "FunctionExpression" &&
-    getFunctionParameters(node.value).length === 0 &&
-    !node.value.returnType &&
-    !isNonEmptyArray(node.value.typeParameters) &&
-    node.value.body
-  ) {
-    return [...(node.decorators || []), node.key, node.value.body];
-  }
-}
-
-/**
- * @param {AstPath} path
- * @returns {boolean}
- */
-function willPrintOwnComments(path /*, options */) {
-  const { node } = path;
-  const parent = path.getParentNode();
-
-  const hasFlowAnnotations = (node) =>
-    hasFlowAnnotationComment(getComments(node, CommentCheckFlags.Leading)) ||
-    hasFlowAnnotationComment(getComments(node, CommentCheckFlags.Trailing));
-
-  return (
-    ((node &&
-      (isJsxNode(node) ||
-        hasFlowShorthandAnnotationComment(node) ||
-        (isCallExpression(parent) && hasFlowAnnotations(node)))) ||
-      (parent &&
-        (parent.type === "JSXSpreadAttribute" ||
-          parent.type === "JSXSpreadChild" ||
-          parent.type === "UnionTypeAnnotation" ||
-          parent.type === "TSUnionType" ||
-          ((parent.type === "ClassDeclaration" ||
-            parent.type === "ClassExpression") &&
-            parent.superClass === node)))) &&
-    (!hasIgnoreComment(path) ||
-      parent.type === "UnionTypeAnnotation" ||
-      parent.type === "TSUnionType")
-  );
-}
+// TODO: Make this default behavior
+const avoidAstMutation = true;
 
 export {
-  handleOwnLineComment,
-  handleEndOfLineComment,
-  handleRemainingComment,
-  getCommentChildNodes,
-  willPrintOwnComments,
+  handleOwnLineComment as ownLine,
+  handleEndOfLineComment as endOfLine,
+  handleRemainingComment as remaining,
+  avoidAstMutation,
 };
